desktop-app/automation-service/requirements.txt
browser-use>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
langchain-google-genai>=2.0.0
websockets>=12.0
pydantic>=2.0.0
